requirements.txt
flask
gunicorn
line-bot-sdk
yfinance
lxml
